refactor: folders reestructuring
