+++ src/main/java/com/bacen/model/BacenEntry.java
@@ -1,39 +1,47 @@
-package com.bank;
+package com.bacen.model;
 
 import java.util.Date;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
-
-/* This is a duplicate class from com.bacen.Entry.
-    In a real world scenario, Bacen and Pixb would be 2 completely different systems, which would transfer data
-    in JSON format. This class is here to simulate that.*/
+import jakarta.persistence.Table;
 
 @Entity
-public class BankEntry {
+@Table(name = "directory")
+public class BacenEntry extends PanacheEntityBase {
     
     @Id
+    @Column(name = "pkey")
     @JsonProperty(value = "pkey")
     public String key;
     
+    @Column(name = "pkey_type")
     @JsonProperty(value = "pkey_type")
     public short keyType;
 
     @JsonProperty(value = "end_to_end_id")
     public String endToEndId;
 
+    @Column(name = "bank")
     public int bank;
 
+    @Column(name = "branch")
     public int branch;
 
+    @Column(name = "account")
     public int account;
 
+    @Column(name = "cpf")
     public String cpf;
 
+    @Column(name = "owner")
     public String owner;
 
+    @Column(name = "creation_date")
     @JsonProperty(value = "creation_date")
     public Date creationDate;
 }
