+++ src/main/java/com/bank/model/Account.java
@@ -1,4 +1,4 @@
-package com.bank;
+package com.bank.model;
 
 import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
 import jakarta.persistence.Column;
