+++ src/main/java/com/bank/model/AccountRegisterDTO.java
@@ -1,4 +1,4 @@
-package com.bank;
+package com.bank.model;
 
 public class AccountRegisterDTO {
     public String cpf;
