+++ src/main/java/com/bank/model/BankPixDTO.java
@@ -1,4 +1,4 @@
-package com.bank;
+package com.bank.model;
 
 public class BankPixDTO {
