+++ src/main/java/com/bank/model/BankPixRequestUpdateDTO.java
@@ -1,4 +1,4 @@
-package com.bank;
+package com.bank.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
