+++ src/main/java/com/bank/model/BankPixRollbacker.java
@@ -1,4 +1,4 @@
-package com.bank;
+package com.bank.model;
 
 public class BankPixRollbacker {
